# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * dancerBase API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileBadge = typeof UserProfileBadge[keyof typeof UserProfileBadge];


export const UserProfileBadge = {
  Newcomer: 'Newcomer',
  Regular: 'Regular',
  Hustler: 'Hustler',
  Legend: 'Legend',
} as const;

export interface UserProfile {
  id: number;
  /** @nullable */
  clerkId?: string | null;
  name: string;
  username: string;
  /** @nullable */
  profilePic?: string | null;
  classesAttended: number;
  badge: UserProfileBadge;
  createdAt?: string;
}

export interface UserProfileUpdate {
  name?: string;
  username?: string;
  profilePic?: string;
}

export interface DanceClass {
  id: number;
  instructorId: number;
  instructorName: string;
  /** @nullable */
  instructorPic?: string | null;
  studioId: number;
  studioName: string;
  studioAddress: string;
  city: string;
  state: string;
  date: string;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  price: number;
  style: string;
  level: string;
  totalSpots: number;
  spotsRemaining: number;
  friendsAttending?: UserProfile[];
  isBooked: boolean;
  /** @nullable */
  bookingId: number | null;
  isSaved: boolean;
  /** @nullable */
  bookingUrl?: string | null;
}

export interface Instructor {
  id: number;
  name: string;
  /** @nullable */
  profilePic?: string | null;
  /** @nullable */
  bio?: string | null;
  styles?: string[];
  upcomingClasses?: DanceClass[];
}

export interface Studio {
  id: number;
  displayName: string;
  address: string;
  city: string;
  state: string;
  upcomingClasses?: DanceClass[];
}

export interface Booking {
  id: number;
  userId: number;
  classId: number;
  class?: DanceClass;
  bookedAt: string;
  confirmationCode: string;
}

export interface BookingInput {
  classId: number;
}

export type FeedItemType = typeof FeedItemType[keyof typeof FeedItemType];


export const FeedItemType = {
  booking: 'booking',
  saved: 'saved',
} as const;

export interface FeedItem {
  id: number;
  type: FeedItemType;
  user: UserProfile;
  class: DanceClass;
  createdAt: string;
}

export interface FriendInput {
  username: string;
}

export interface PastClass {
  id: number;
  classId: number;
  class?: DanceClass;
  /** @nullable */
  songPlayed?: string | null;
  videoLinks?: string[];
  attendees?: UserProfile[];
  attendeeCount: number;
  attendedAt: string;
}

export type SearchUsersParams = {
q?: string;
};

export type ListClassesParams = {
search?: string;
city?: string;
style?: string;
level?: string;
instructor?: string;
dateFrom?: string;
dateTo?: string;
priceMin?: number;
priceMax?: number;
timeFrom?: string;
timeTo?: string;
sort?: ListClassesSort;
};

export type ListClassesSort = typeof ListClassesSort[keyof typeof ListClassesSort];


export const ListClassesSort = {
  trending: 'trending',
  date: 'date',
  price: 'price',
} as const;