# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: dancerBase API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List and search classes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: instructor
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: priceMin
          in: query
          schema:
            type: number
        - name: priceMax
          in: query
          schema:
            type: number
        - name: timeFrom
          in: query
          schema:
            type: string
        - name: timeTo
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending, date, price]
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DanceClass"
  /classes/{classId}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get a single class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dance class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DanceClass"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: Get current user's upcoming bookings
      responses:
        "200":
          description: Upcoming booked classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking receipt
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
...
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List current user's friends
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: addFriend
      tags: [friends]
      summary: Add a friend by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendInput"
      responses:
        "200":
          description: Friend added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /friends/{userId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
...
  /classes/{classId}/friends-attending:
    get:
      operationId: getFriendsAttending
      tags: [classes]
      summary: Get friends attending a class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friends attending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: ["string", "null"]
        name:
          type: string
        username:
          type: string
        profilePic:
          type: ["string", "null"]
        classesAttended:
          type: integer
        badge:
          type: string
          enum: [Newcomer, Regular, Hustler, Legend]
        createdAt:
          type: string
      required:
        - id
        - name
        - username
        - classesAttended
        - badge
...
    DanceClass:
      type: object
      properties:
        id:
          type: integer
        instructorId:
          type: integer
        instructorName:
          type: string
        instructorPic:
          type: ["string", "null"]
        studioId:
          type: integer
        studioName:
          type: string
        studioAddress:
          type: string
        city:
          type: string
        state:
          type: string
        date:
          type: string
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        price:
          type: number
        style:
          type: string
        level:
          type: string
        totalSpots:
          type: integer
        spotsRemaining:
          type: integer
        friendsAttending:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        isBooked:
          type: boolean
        bookingId:
          type: ["integer", "null"]
        isSaved:
          type: boolean
        bookingUrl:
          type: ["string", "null"]
      required:
        - id
        - instructorId
        - instructorName
        - studioId
        - studioName
        - studioAddress
        - city
        - state
        - date
        - dayOfWeek
        - startTime
        - endTime
        - price
        - style
        - level
        - totalSpots
        - spotsRemaining
        - isBooked
        - bookingId
        - isSaved
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        classId:
          type: integer
        class:
          $ref: "#/components/schemas/DanceClass"
        bookedAt:
          type: string
        confirmationCode:
          type: string
      required:
        - id
        - userId
        - classId
        - bookedAt
        - confirmationCode
    BookingInput:
      type: object
      properties:
        classId:
          type: integer
      required:
        - classId
...
    FriendInput:
      type: object
      properties:
        username:
          type: string
      required:
        - username
...