# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * dancerBase API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DanceClass } from './danceClass';

export interface Booking {
  id: number;
  userId: number;
  classId: number;
  class?: DanceClass;
  bookedAt: string;
  confirmationCode: string;
}