# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * dancerBase API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileBadge } from './userProfileBadge';

export interface UserProfile {
  id: number;
  /** @nullable */
  clerkId?: string | null;
  name: string;
  username: string;
  /** @nullable */
  profilePic?: string | null;
  classesAttended: number;
  badge: UserProfileBadge;
  createdAt?: string;
}